Moved C-macro parsing logic to op2_gen_common.py
